+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.4.0
 
 - Remove files from Edit Mods Dialog that don't exists anymore
+- Add Show In File Explorer button to Edit Mods Dialog right click menu